+++ packages/utils/expand-transition.js
@@ -1,6 +1,7 @@
 import { css, html } from 'lit';
 import { collapse, expand } from 'element-collapse';
 import { FabricElement } from '.';
+import { ifDefined } from 'lit/directives/if-defined.js';
 
 class ExpandTransition extends FabricElement {
   static properties = {
@@ -26,6 +27,10 @@ class ExpandTransition extends FabricElement {
     }
   }
 
+  firstUpdated() {
+    this._mounted = true;
+  }
+
   updated() {
     if (!this._wrapper) return;
 
@@ -36,10 +41,6 @@ class ExpandTransition extends FabricElement {
 
     if (this._mounted && !this.show && !this._removeElement) {
       collapse(this._wrapper, () => (this._removeElement = true));
-    }
-
-    if (!this._mounted) {
-      this._mounted = true;
     }
   }
 
@@ -54,7 +55,7 @@ class ExpandTransition extends FabricElement {
   `;
 
   render() {
-    return html`<div aria-hidden=${!this.show}>
+    return html`<div aria-hidden=${ifDefined(!this.show ? 'true' : undefined)}>
       ${this._removeElement ? html`` : html`<slot></slot>`}
     </div>`;
   }