+++ vite.config.js
@@ -80,6 +80,11 @@ export default ({ mode }) => {
             injectOptions,
           },
           {
+            filename: 'textfield.html',
+            template: 'pages/components/textfield.html',
+            injectOptions,
+          },
+          {
             filename: 'index.html',
             template: 'index.html',
             injectOptions,